+++ TPie/Config/RingSettingsWindow.cs
@@ -71,7 +71,7 @@ namespace TPie.Config
                     Ring.Color = new Vector4(color.X, color.Y, color.Z, 1);
                 }
 
-                if (Ring.KeyBind.Draw(Ring.Name))
+                if (Ring.KeyBind.Draw(Ring.Name, 310 * _scale, true))
                 {
                     Plugin.Settings.ValidateKeyBind(Ring);
                 }